+++ packages/core/src/components/TextField/Styled/TextField.styled.tsx
@@ -3,10 +3,22 @@ import { StyledProps } from '../types';
 import { Label } from './Label.styled';
 import { MaskPlaceholder } from './MaskPlaceholder.styled';
 
+const transformLabel = (variant: string) => {
+    if (variant === 'fusion') {
+        return css`
+            transform: translateY(-140%) scale(0.67);
+            background-color: white;
+            padding: 0 0.4rem;
+            z-index: 1;
+        `;
+    }
+    return 'transform: translateY(-87%) scale(0.75)';
+};
+
 const styleWithLabel = ({ variant }: StyledProps) => {
     return css`
         align-self: flex-end;
-        padding-bottom: ${variant === 'filled' ? '0.7rem' : '0.6rem'};
+        padding-bottom: ${variant === 'filled' ? '0.7rem' : variant === 'fusion' ? '1.4rem' : '0.6rem'};
     `;
 };
 
@@ -20,6 +32,7 @@ export const Input = styled('input')<StyledProps>`
     background: transparent;
     border: none;
     text-overflow: ellipsis;
+    z-index: 1;
 
     ${({ isLabelPresent, size }) => isLabelPresent && size !== 'S' && styleWithLabel};
     &,
@@ -47,7 +60,7 @@ export const Input = styled('input')<StyledProps>`
 
     &:not(:placeholder-shown) ~ ${Label}, &:focus ~ ${Label} {
         opacity: ${({ size }) => size === 'S' && '0'};
-        transform: ${({ size }) => size === 'M' && 'translateY(-87%) scale(0.75)'};
+        ${({ size, variant }) => size === 'M' && transformLabel(variant)};
     }
 
     &:not(:placeholder-shown) ~ ${MaskPlaceholder} {