+++ packages/core/src/components/TextField/types.ts
@@ -3,7 +3,7 @@ import { HTMLProps, Omit, WithThemeProp } from '@medly-components/utils';
 
 export interface TextFieldProps extends Omit<HTMLProps<HTMLInputElement>, 'prefix' | 'size' | 'height'>, WithThemeProp {
     /** Input Variants */
-    variant?: 'outlined' | 'filled';
+    variant?: 'outlined' | 'filled' | 'fusion';
     /** Input Size */
     size?: 'S' | 'M';
     /** Function will called with the input value on blur and invalid event */
@@ -38,12 +38,14 @@ export interface StyledProps extends TextFieldProps, TextFieldTheme {
     isPrefixPresent: boolean;
     isSuffixPresent: boolean;
     isLabelPresent: boolean;
+    variant: 'filled' | 'outlined' | 'fusion';
 }
 
 export interface InnerWrapperProps extends Omit<HTMLProps<HTMLDivElement>, 'size' | 'height'>, TextFieldTheme, WithThemeProp {
-    variant?: 'outlined' | 'filled';
+    variant?: 'outlined' | 'filled' | 'fusion';
     disabled?: boolean;
     size?: 'S' | 'M';
     isLabelPresent?: boolean;
     isErrorPresent?: boolean;
+    isTextPresent?: boolean;
 }