+++ packages/core/CHANGELOG.md
@@ -3,6 +3,56 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.28.2](https://github.com/medly/medly-components/compare/@medly-components/core@1.28.1...@medly-components/core@1.28.2) (2020-10-14)
+
+
+### Bug Fixes
+
+* **core:** fix pagination page selector popup placement and set pag button's type attribute to butto ([2b441b0](https://github.com/medly/medly-components/commit/2b441b086943bc3767c734995f22e401e01ae3be))
+
+
+
+
+
+## [1.28.1](https://github.com/medly/medly-components/compare/@medly-components/core@1.28.0...@medly-components/core@1.28.1) (2020-10-14)
+
+
+### Bug Fixes
+
+* button type issue in calendar component ([2a2f657](https://github.com/medly/medly-components/commit/2a2f657dca2b2186a023c41674dc33f8e764d5ef))
+
+
+
+
+
+# [1.28.0](https://github.com/medly/medly-components/compare/@medly-components/core@1.27.1...@medly-components/core@1.28.0) (2020-10-13)
+
+
+### Features
+
+* add new input variant to theme, add optional text and remove shadow for new variant ([980fd03](https://github.com/medly/medly-components/commit/980fd03965b423672e0512c6b1ff383954e873d8))
+* change inputContainsText to isTextPresent ([59d19cb](https://github.com/medly/medly-components/commit/59d19cb4d1ee3851b1da81cf8de0513517fdd1f8))
+* change name of new variant from rounded to outlinedRounded ([d04f844](https://github.com/medly/medly-components/commit/d04f844c2b55c94f146fdc3dd5e553160ffb16e3))
+* change new variant name from outlinedRounded to fusion ([1db2ad2](https://github.com/medly/medly-components/commit/1db2ad29a9b31bea061165a527706afd6c62535a))
+* fix failing tests, update snapshots ([01452e3](https://github.com/medly/medly-components/commit/01452e373306bdad78e2c3c2644cc8c1ea161bfa))
+* reorganize inner wrapper styles, fix error style bug on filled variant ([9beac01](https://github.com/medly/medly-components/commit/9beac0121086b5ca1b151b6cfbedd1440904caf6))
+* update snapshots ([4373fe0](https://github.com/medly/medly-components/commit/4373fe08e5af6f607957716dd3dd3ac060e2a7cc))
+
+
+
+
+
+## [1.27.1](https://github.com/medly/medly-components/compare/@medly-components/core@1.27.0...@medly-components/core@1.27.1) (2020-10-08)
+
+
+### Bug Fixes
+
+* memoized onChange method issue in textField component ([9f8301e](https://github.com/medly/medly-components/commit/9f8301eabe3032b7715e68952b687ef362b46899))
+
+
+
+
+
 # [1.27.0](https://github.com/medly/medly-components/compare/@medly-components/core@1.26.0...@medly-components/core@1.27.0) (2020-10-08)