+++ packages/loaders/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.27](https://github.com/medly/medly-components/compare/@medly-components/loaders@1.0.26...@medly-components/loaders@1.0.27) (2020-10-13)
+
+**Note:** Version bump only for package @medly-components/loaders
+
+
+
+
+
 ## [1.0.26](https://github.com/medly/medly-components/compare/@medly-components/loaders@1.0.25...@medly-components/loaders@1.0.26) (2020-10-08)
 
 **Note:** Version bump only for package @medly-components/loaders